Made the tests compile and pass. Still need to work on coverage because I probably don't cover 50% at the moment (I added a lot of stuff!)
Use the getopt_exception_logic whenever a problem is found in tables/calls (i.e. programmer has to fix it before releasing his software.)
Added a get_options() in the getopt class to be able to handle names of 1 character and get the option using the short name instead.
Changed the default value of a standalone flag to be an empty string which is enough for the value to be defined (and quite different from using "1".)
Fixed the logger so it actually works. The templates have to be in the header and we need a specialization for the functions (end() and nobr()).
Fixed the clearing of the f_log stringstream by setting it to an empty string and not calling clear() anymore.
Fixed the parse_options_info() function, it would not pass through the first test where I used && instead of ||.
Added support for the "i" flag in "/<regex>/i" for validators.
Added support for the validator specification in the option definition file to have a parameter (i.e. "validator=regex(/my-regex/)")
Added call the add_child() when done parsing the option from a file.
Fixed all the tests of the GETOPT_FLAG_COMMAND_LINE and GETOPT_FLAG_ENVIRONMENT_VARIABLE, they were inverted (!).
Fixed error messages that said "tool" by "program" instead.
Added some missing "<< end" for various log calls.
Fixed several add_option() calls (some were duplicated, one was calling add_options() instead.)
Allow calling process_help_string() with a nullptr.
Added a GETOPT_FLAG_DEFAULT_OPTION so the default option can be given a name other then "--".
Fixed the add_value()/set_value() when necessary (idx == size()) do a push_back() otherwise it crashes.
Fixed the get_long(), the "end" variable would be confused with the log "end()" function!
Fixed the split_string(), I meant to use boost::trim() (not trim_copy).
Properly push the default option to the very end of the usage() screen (still before the help_footer).
Added support for '\r' in configuration files.
Added the validator implementation (register & create functions)
Added a regex validator.
Added the integer and regex validator factories.
Changed the getopt() constructor so it accepts a reference to an options_environment instead of a pointer.
Added a "-t" to run tests from my ./mk script.