Fixed the version. It now properly makes use of the SnapGetVersion() function.
Removed the getopt_exception_exiting exception. We do not use it anymore.
Removed the nobr() in the logger. I'm not too sure why I had that there. it's a usage() _feature_.
Enhanced the split_string() to support quoted strings. When a quoted string appears, anything is accepted between the quotes.
Bumped version to 2.0.0.0 because it's really a big change even if mostly compatible with the old version.
Adding coverage script to make sure we test most everything.
Added tests of the version functions.