Made the path to the file of option declarations changeable.
Fixed the handling of sections in the config_file.cpp.
Use the list of sections in the parse_options_from_file() function.
Unquote the default value just so we allow quoting of it.
Fixed the minimu/maximum set_...() calls in parse_options_from_file().
Made the GETOPT_FLAG_SHOW_ALL the default of usage().
Fixed the test of the first character in a config file, we need to replace '_' to '-' first.
Added a new test to check options in an option .ini file.