Added SnapDev as a dependency. I use NOTUSED() already.
Prevent printing the footer if the string is empty. That way we can add a newline just before it.
Fixed the `breakup_line()` in case the input string has a newline.
Moved the option class and many new functions to its own header.
Fixed the type of the `short_name_t` to a `char32_t`.
Added an `debian/source/options` file to avoid the .git and tmp folders in our source files.
Aded a `build-file-of-options` tool to transform .conf comments into configuration definition files.
Looked at the converting `hide-warnings` to C++ but it's a bit more than I first thought.