Added a new test: Option Parser to specifically test the parsing of the options table.
Added the ability to specify the name of a test with the ./mk -t command.
Fixed the system options, they needed to use the standalone_command_flags().
Added a test of argv[0] in case it is nullptr and thus no program name is available.
Prevent default options to include the FLAG flag.
Prevent an alias from having a help string (in an option file.)
Changed the alias name location. It is expected to be in the help string, not the default.
Fixed the set_alias(), we have to test the ALIAS flag from the alias object.
Removed the default value from the OptionValue. It's not necessary.
Added an alias so it's possible to verify that we are defining an alias instead of a help string and prevent/allow certain things.
Allow specifying a test name on the dev/coverage command line so we can see that one test does what it is expected to do.