Broke up the advgetopt.cpp file in 6 files so it's easier to write tests that match each file instead of wondering whether the coverage is indeed complete.
Changed the getoption() so it returns the final option in case it is an alias. That way whether you use the name of the alias or aliased option you get the expected results (i.e. the is_defined() and get_string() functions show you the option that receives the user value.)
Added an add_option_from_string() function to support the long options with an equal sign. The value(s) come in one string.
Prvent the creation of the default option ("--") with a short name.
Renamed the get/set alias functions as get/set alias destination since the pointer is not the alias option but the destinatio nof this alias.
Cleaned up the Separators declarations (names and constness).
Added a tests/README.md where I can list the name of the tests to run to verify coverage.
Added a --verbose to the tests so the log_for_test() uses --verbose instead of the --progress.
Fixed the test of exceptions with our new ExceptionMatcher class (instead of having an std::string operator in our exceptions.)