Added the '%e' and '%*e' to print out the environment variable name and value.
Changed the default usage show flags with GETOPT_FLAG_SHOW_MOST.
In case of a very long argument name, make sure to have an indentation for the first line of help.
In the breakup_line() remove spaces in the line after we break it up (i.e. if we cut the line in a place with 2 or more spaces, we remove all those spaces).
Changed the combine_option_flags() first assert. We do not need any of these flags to be set if there is the command line, environment variable, or configuration file flags which is checked in the next assert.
Added the usage.cpp test which verifies 100% of the usage() and subfunctions used by usage().