Added many more tests (logger, option, utilities, validators, version)
Fixed the name of the test in main.cpp (copy/paste mistake)
Updated a few tests that required it as I made changes.
Changed the parser to use add_flags() for options, instead of set_flags().
Added an has_default() function, because get_default() may return an empty string as the default.
Removed the load_configuration_files(). We now use the insert_project_name() and the get_configuration_filenames().
Added a get_configuration_filenames() which returns the list of all the filenames the loader will check.
Allow for the "~" again, it gets transformed with the contents of the $HOME environment variable.
Removed the minimum=... maximum=... from configuration definitions. Instead these go in a validator if supported by said validator.
Make sure that the validator factories get instantiated.
Added a parser in the validator_integer() constructor to allow for range and specific numbers to be accetable (min/max. replacement.)
Redesign the validator_integer::validate() function to take our new convert_string() function and the ranges in account.
Properly test for overflows in the validator_integer::convert_string() function.
Fixed up the validator_regex constructor so it still accepts a regex where the ending '/' is missing (and it does not pass an empty string.)
Moved the flags definitions and functions to a separate header.
Added tests to allow "--" as an argument name but prevent other names from starting with a dash (-).
Prevent the short name from being '-'.
Verify the index in option_info::validates().
Fixed various exceptions (i.e. use invalid instead of unknow, or logic instead of anything else.)
Changed the option_info::add_value() to make use of the set_value() so we benefit from all the tests (MULTIPLE vs single cases especially.)
Added a call to the validates() function whenever a new value is added/set.
Enhanced the set_multiple_value(), if used with a single value the result must be a single value or it throws a logic error.
If we have a validator, verify the list of values after the set_multilpe_value() parsed the input. Remove invalid entries.
Make the get_long() use our validator_integer::convert_string() which we have full control over.
Fixed the split_string() quotation checks. A separator is not required before/after a quote now.
Added an insert_project_name() and handle_user_director() set of utilities.
Fixed some documentation/comments.