Fixed a bug in the set_multiple_value(), it would not clear the f_integer variable member.
Added an has_flag() function to test the environment flags.
Removed the add_options() now that the add_value() is fixed (i.e. does not throw for trying to add in a non-multiple.)
Fixed the case where a flag is specified but not its value (and it's not a REQUIRED flag), then we set the value to the empty string which is important because the default value is not adequat in many cases in such a circumstance.
Throw an exception when the name is empty calling the [] operator, including the const version.
If calling the non-const [] operator and the value is not yet defined, define it as an empty string.
Added a string in the options so we can have validators there too. Sine we do not have the min/max on the get anymore, we need to have a fallback.
Moved the parsing of the validator parameters inside a validator::create() function which allows us to use that feature directly with validators. Much better.
Allow for setting the validator as a string now.
Changed the parameter data to the validator constructors to a list of strings which represent the parsed parameters.
Fixed a bug in the configuration file parser which would break the current section name using the '{' and '}' brackets.
Deleted unused test logic_errors.cpp (it's full replicated now.)
Merged and renamed the valid_config_*.cpp files. It's now in config.cpp and runs against the advgetopt_config.cpp, not the config_file.cpp.
Fixed variable documentation snippets.
Made changes to the README.md, especially add the Option Info References entry and the table (Library vs Tests.)
Added tests to cover all the little additions I've made lately. All the tests are 100% coverage again.
Added -j4 for a couple of make in ./mk.