Added the config_file.cpp test. It is coverage at 69.4% so far.
Added a dependency on pthread because we want the configuration files to be accessible from any thread and thus need a mutex.
Changed my mind and removed all the parent/child concepts. Instead we just have a list of sections incorporated in the name of each option (as before). Much less work to access a parameter although it always has to be fully qualified.
Moved the short name conversions to the option_info.h/cpp instead of attempting conversions everywhere.
The conversion from a name in a string to a short name accepts any string as input, if the input string is not exactly one character, the function returns NO_SHORT_NAME every time.
Added a separate setup for configuration files. Right now the constructor initializes that setup, but later we may offer functions for extensions.
Implemented the configuration file cache (when the pthread dependency). This means any one file gets loaded once per run of your app. You can't directly create a configuration file, you have to do a get_conf_file() which checks the cache first.
Made the configuration support an URL (output only for now). This is useful to make sure that when requesting the same file twice we use the exact same parameters.
Actually test the SECTION flags on loading of a configuration file. So if FILE_INI is not set, the '[' and ']' have no special meaning.
Prevent the creation of configuration files without a filename (the file may not exist, the filename may not even be valid, but it has to not be empty.)
Finally wrote a set_parameter() function. It verifies the section information. We now transform the '[' ... ']', '{' ... '}' and '.', to '::' so once everything is parsed, you have a single section separator.
Fixed the conf_file::getc(), the f_unget_char was not properly handled.
Fixed the ungetc() function, I was comparing to '\0' with the wrong operator (i.e. '==' instead of '!='.)
Fixed several line continuations, they would not get the next character (skipped the continuation character, but did not read the next character after that.)
In case of the semicolon support, I now check the first line for a comment, if it is a comment, it gets removed. It's not perfect... each newline should allow for a new comment.
Changed the list of questions to a set instead of a map. We do not need a value other than the key.
Added masks of possible flags in the configuration setup. This especially helps with the tests.
Added an initialization exception so we can specifically throw on such failure. At this point it gets used in case the pthread mutex initialization fails.
Changed the validation process. Any time you set a value or change the validator, values get validated. An invalid value is rejected so it never remains in the list of values.
Many functions now return true when a value gets added and false when it does not (i.e. it is considered invalid.)
The low level validates() function now takes care of all checks, if the value is empty, there is no validator, and a few other things.
Fixed the coverage script, the gcov was accessing the wrong folder.
Complete the converage tests for the advgetopt_config.cpp file.
Updated many of the tests accordingly. Some of the changes required updates so it compiles, runs, and covers all new code (deleted a few tests too since some code was deleted and some is not accessible anymore.)
Corrected a few messages.