Removed what is done from the TODO.txt.
Added a missing const to a function.
Updated documentation about the supported % in the process_help_string().
Added support for the %[*]g, %o and %w percent flags. Added the corresponding tests.
The dev/coverage command of the tests/README.md for the conf_file.cpp had a missing test in the list.
Added a some usage tests so we can test all the new percent flags.
Added tests for the case where an option is or not an alias and we want to grab the alias option, not the destination.
Enhance some tests which verify that the short and long names are indeed returning the same option.
Moved the init_tmp_dir() to the main.cpp/.h because that way it can be used anywhere (we now use it in a third file.)
Fixed several tests that used hard coded paths that were specific to my installation.