Added the function to actually process system options. We do not want each user to have to rewrite that each time.
Added a system --config-dir option. It can be used any number of times.
Handle Unicode command line options (short names where still ISO-8859-1).
Allow for setting a short name to an option which does not have one yet.
Transformed the option_info_ref to support a reference to a still undefined option (i.e. it does not set it until an assignment operator is used.)
Fixed many Unicode characters, they are expected to be defined with U'...' and not L'...'.
Moved the default options files under `/usr/share/advgetopt/options` allowing us to have other folders there later.
Added the %i to the help strings, it shows us the directory where to read files of options.
Added support for char32_t characters in the option reference.
Added the bool and ! operators to the option reference.
Added a get_long() to the option reference.
Added string + operators to be able to add char32_t to a string. This may need to go to libutf8, but we have a special case for '\0'.
Added missing documentation to the option reference class.
Updated the test/README.md with the new tests added to support all the new code.
Updated many tests so they work with the new code.
Added new tests so the coverage stays at 100%.
Updated some documentation.