Put the group descriptions to the options_environment.
Put system arguments in the COMMANDS group (which is group ONE renamed.)
Put the --config-dir to the OPTIONS group (which is group TWO renamed.)
Added a flag to have system parameters automatically processed by the constructor.
(Re-)Added the getopt_exception_exit so if a system parameter representing a command is executed, we throw that exception. This is much better for the library than an exit().
Added code to retrieve the console line width to write the usage().
Fixed the --environment-variable-name output, if not defined, write a line saying so.
Implemented a function to create options from the name of groups (`--<group-name>-help`).
Write/take groups in account only if the f_groups is defined (not nullptr.)
Added templates to initialize the `group_description` which replaces the set_group_name().
Added a couple of tests to cover the new additions.
Updated a few more tests so they pass with the new changes.
Created a new macro to find differences in long strings when not equal.