Moved the log.cpp/h from advgetopt to cppthread so we can depend on cppthread (instead of the other way around). This gives us access to the cppthread::mutex object.
Added cppthread as a dependency.
Made insert_project_name() return a list of filenames since we now support the `??-<name>.conf` format for user filenames.
Renamed the exceptions as errors and use the libexcept macros.
Cleaned up the #include in many places.
Fixed the namespace comment (closing the namespace).
Cleaned up the copyright notices.
Fixed a couple of files where the copyright notice was wrong.
Updated the tests accordingly.