Fixed bug when the configuration filename is a directory.
The getline() function would never return if the input file was a directory. So I added code to test and if it is a directory, we generate an error and return immediately. Since we generally ignore errors while reading configuration files, this is going to be transparent. The process will attempt to log the problem, but of course, the snaplogger gets initialized just after the logger configuration files get loaded...