Fixed the tests so they actually pass with the newest version.
Addeda group name field so we can have a common sub-directory to many tools (i.e. "snapwebsites.d").
Changed the insert_project_name() into insert_group_name().
Added a %s to print out the group name in the usage, etc.
Fixed the configuration getc(), we now fail on "!in".
Do not add the "\n" to the lines read from the configuration file unless it is inside the value (i.e. `trim()`).
Configuration files are not required to exist. If missing it just makes it as ready.
Moved the EOF test after the while tool.
Chagned the EOF -> invalid to valid. Now we accept non-existant files and it works. (This is useful when you want to create a new file).
Added a few tests to increase coverage %.
Various small clean ups and fixes (i.e. missing character).