Added --show-option-sources feature.
Fixed a bug in the configuration file directory discovery, I was trying to use the "--config-dir" too early.
Added a flag to know whether the parsing is done or not and throw if certain functions are called beforethe flag is true.
Fixed some error logs that would reference the environment variable when the error was about command line.
Fixed a bug with several system options. If the corresponding pointers were not set, the library would crash.
Added '-?' as the short name of the '--long-help' option.
Added a callback capability on the option_info.
Enhanced the callback of the config_file to support any number of functions.
Added a DYNAMIC flag to allow, or not, the set_value() on getopt objects.
Ordered the system options.
Added quotes around some of the command line options in error messages.
Made the "set_multiple_values()" plural.
Prevent changes to object fields in "set_multiple_values()" on most errors.