Added the edit-config tool to edit Unix like configuration files.
Allow for saving the comments when reading configuration files.
Allow for writing the comments back out in save_configuration().
Changed the backup flag in a string so we can specify the extension.
Allow to save in a file other than the file we loaded from.
Create sub-directories if necessary to save the new configuration file.
Replaced the boost replace_all() calls with one string_replace_many().
Fixed a bug in the get_line() while reading the last line without a '\n'.
Return default value on a get_string() when option is used by itself and REQUIRED flag is not set.