More support to accept duplicates.
We want the fluid-settings to be able to reload the configuration files
because we may get some of these "late" (i.e. copied from another computer
to the fluid-settings computer). So we need to support duplicates because
when we re-load it will reload the same files and thus find duplicates.
All the same options are going to be found twice except the few new ones
found in the new file.