Fixed backup filename.
Fixed the set_comment(), if empty after a trim, clear the comment.
Fixed the get_comment() in case we print it, we want to make sure there is a final `\n` otherwise it will be merged with its value.
In the save_configuration, make sure to use the `output_filename` if defined (we have to backup the file which is about to be smashed).
Print out the warning to file even if the list of parameters is empty, only avoid if the first parameter has a comment.
Added a note about the separators which are not applied in all cases (although they probably should be).
Cleaned up the cmake file (AdvGetOptConfig.cmake).
Added comments in the AtomicNamesConfig.cmake about why we have two find_program()'s,
Updated a few tests so they pass with the new --print-option and an error which now prints the line number.