Got validator tests covering 100%.
This version enhance the validator test to cover 100% of the validators.
I made some clean ups (better error messages).
Forbid using two periods as in: "2..8 years" because that meant "2.0 seconds
and 0.8 years" (which is was probably a mistake from the user who meant
"2.8 years" or "2.0 seconds ... 0.8 years" as a range).
Moved the power128() function to our snapdev library.
Broken up the invalid validator tests in separate tests before adding the
duration, size, and double invalid tests. Cleaner!