Cleaned up the configuration filename generation.
Made the options filename a public function: get_options_filename().
It was used in two places, so to make sure it stays in sync., in a function works better.
Put the group or project name retrieval in a function so it's always the same everywhere.
Allow parse_options_from_file() to be called with an empty filename.
Fixed the tokenization of the section names in parse_options_from_file(), it would see each ':' as separate symbol.
Allow for an extract section to be ignored in the options file. That way we can load fluid settings as options.
Added support for the "dynamic-configuration" flat in "allow=...".
Added support for a "...::group" parameter to place the options in a group (i.e. COMMANDS, OPTIONS, etc.).
Show dynamic options with a "*" at the end of the name/arg in the help screen.