Fixed the output filename and default paths.
The code changed a lot and the default output filename which was gathered
from the complete list of files was not really coming out properly. We now
have a separate function to define that path properly every time (at least
the algorithm is clear now and it takes the priority in account as
expected--the old code would take the last file which was wrong now
that we have a priority).

Applied a small optimization to the search of variable sections.

Fixed the function name in an exception error message.

Added an unset to the ./mk script because as a programmer I'm likely to have
the ADVGETOPT_OPTIONS_FILES_DIRECTORY defined so it can find locally installed
option files (very much used by fluid-settings).

Updated the tests to match the new configuration filename order and the fact
that the system defined filenames are generated in many more cases now
(i.e. %i, %o and %*g are quite different now).