Added functions to convert options to/from '-'/'_'.
The library used boost to replace the '-' in '_' and vice versa.
There are external users that may be interested in doing the same conversions.
For that reason, I added two utils functions to do those conversions.
I updated the library so it uses those functions instead of boost.
I removed the #include of the boost library wherever possible so far.
(I still have some join() and trim() which are available in snapdev, I just
don't want to convert the code at the moment... will happen soon).