Show aliases along aliased option.
I added a --dry-run command line option to the ipload tool which also has
the --verify command. The dry run is an alias and it would not show up in
the usage output. This newer version adds aliases to the list of options
(i.e. --verify or --dry-run or -V).