Try fixing issue on build system.
Somehow, the build fails on the basic_string + operation with an overflow.
It's "just a warning" but I still think that it should not warn like that.
There seem to be posts about it on the g++ bugzilla so I don't really need
to do more here. I need my build to work, though.